Refresh current conditions on a given time interval

+++ WeatherMenu/API/WWO/WWOApiClient.h
@@ -12,7 +12,7 @@
 
 + (WWOApiClient *)instance;
 
-- (void)currentConditionsIn:(NSString *)city forDays:(int)daysNo responseHandler:(void (^)(NSDictionary *))success;
+- (void)currentConditionsIn:(NSString *)city forDays:(NSInteger)daysNo responseHandler:(void (^)(NSDictionary *))success;
 
 
 @end
